ShipGameData

+++ Assets/_ChromaCrusade/Scripts/Data/ShipPart/ShipCabinData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NewShipPart", menuName = "Data/Ship Cabin")]
 public class ShipCabinData : ShipPartData
 {
-    public float handling;
+    public int handling;
 
     public override PartType PartType => PartType.Cabin;
 }
