+++ Assets/_ChromaCrusade/Scripts/Data/ShipPart/ShipCoreData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NewShipPart", menuName = "Data/Ship Core")]
 public class ShipCoreData : ShipPartData
 {
-    public float energy;
+    public int energy;
 
     public override PartType PartType => PartType.Core;
