+++ Assets/_ChromaCrusade/Scripts/Data/ShipPart/ShipPartData.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class ShipPartData : ScriptableObject
 {
     public Sprite sprite;
-    public float mass;
+    public int mass;
     public int price;
     public virtual PartType PartType { get; }
